Separated Minimax function from CPUChoice class and created a new Minimax class
